Classes have been added in Adapters, an abstract class and a derived class for Excel, which will accept a Template. A class has also been created for Encryption and Decryption of information. Classes have been created for data access in Databases, both in Microsoft's SQL Server, as well as for SQLite and FireBird. Have added Extension classes to the language, the Enumerated Interface and for the Lists Interface, as well as for the Integer, Date String and Array types.
All of this has, except for the extensions, a Test class in the SIE.Core.Test project.

A project called SIE.Core.Tests has been created, which is a library dependent on SIE.Core, which has the necessary classes to develop the other TEST libraries of the Integral Business System (SIE).